requirements.txt
numpy==1.25.0
pandas==2.1.1
google-cloud-bigquery==3.14.0
google-auth==2.23.0
pyarrow==13.0.0